chore: align node 24 config and tooling
