+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
       - uses: pnpm/action-setup@v4
       - uses: actions/setup-node@v4
         with:
-          node-version: 20
+          node-version-file: .nvmrc
           cache: pnpm
       - run: pnpm i -w
       - name: Commitlint (PR commits)
@@ -49,5 +49,5 @@ jobs:
       - run: pnpm lint
       - run: pnpm typecheck
       - name: SSR guard
-        run: node scripts/assert-no-client.mjs
+        run: node scripts/assert-no-client.js
       - run: pnpm build
