+++ apps/speckit/next.config.js
@@ -1,0 +1,5 @@
+const { defineMonorepoNextConfig } = require("../../next.shared-config.js");
+
+const config = defineMonorepoNextConfig(__dirname);
+
+module.exports = config;
