--- apps/speckit/next.config.mjs
@@ -1,5 +1,0 @@
-import { defineMonorepoNextConfig } from "../../next.shared-config.mjs";
-
-const config = defineMonorepoNextConfig(import.meta);
-
-export default config;
