+++ apps/speckit/package.json
@@ -11,7 +11,7 @@
     "lint": "next lint",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "a11y": "pa11y-ci || echo 'skip'",
-    "links": "node ../../scripts/check-links.mjs",
+    "links": "node ../../scripts/check-links.js",
     "sbom": "echo 'generate sbom'"
   },
   "dependencies": {
@@ -29,7 +29,7 @@
     "@tailwindcss/postcss": "^4.1.13",
     "@tailwindcss/typography": "^0.5.19",
     "@types/node": "^24.5.2",
-    "@types/react": "^19.1.14",
+    "@types/react": "^19.1.15",
     "@types/react-dom": "^19.1.9",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.36.0",
